construction/scripts/requirements-3.10.txt
cbor2==5.4.6
et-xmlfile==1.1.0
Levenshtein==0.20.9
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.3
pycryptodome==3.16.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
rapidfuzz==2.13.7
six==1.16.0
